backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
python-multipart
